+++ R/dates.R
@@ -3,11 +3,14 @@
 #' Helper functions for manipulating and converting between regular year-month-day date strings and EIA date string notation.
 #'
 #' There is no reason to mix EIA date formats in this context. Functions that take EIA date strings expect a consistent format.
+#' Also, EIA date formats are parsed automatically from the dates themselves.
+#' However, daily and weekly use the same format. Too avoid ambiguity in \code{eia_date_seq}, daily is assumed; set \code{weekly = TRUE} to treat as weekly.
 #'
-#' @param x character, EIA date string; character or date object for for regular dates. Dates should be in year-month-day order.
+#' @param x character, EIA date string; character or date object for for regular dates. Dates should be in \code{yyyy-mm-dd} format.
 #' @param start start EIA date or date.
 #' @param end end EIA date or date.
-#' @param date_format EIA date format: "A", "Q", "M". These stand for annual, quarterly, monthly.
+#' @param date_format EIA date format: "A", "Q", "M", "W", "D". These stand for annual, quarterly, monthly, weekly, daily.
+#' @param weekly logical. See details.
 #'
 #' @export
 #' @name eiadate
@@ -24,30 +27,32 @@
 eiadate_to_date <- function(x){
   date_format <- eiadate_format(x)
   switch(date_format,
+         "A" = lubridate::ymd(x, truncated = 2L),
          "Q" = lubridate::yq(x),
          "M" = lubridate::ymd(x, truncated = 1L),
-         "A" = lubridate::ymd(x, truncated = 2L))
+         "D" = lubridate::ymd(x))
 }
 
 #' @export
 #' @rdname eiadate
-date_to_eiadate <- function(x, date_format = c("A", "Q", "M")){
+date_to_eiadate <- function(x, date_format = c("A", "Q", "M", "W", "D")){
+  x <- lubridate::ymd(x)
   date_format <- match.arg(date_format)
-  if(date_format == "Q"){
+  if(date_format == "A"){
+    substr(as.character(x), 1, 4)
+  } else if(date_format == "Q"){
     gsub("\\.", "Q", lubridate::quarter(x, with_year = TRUE))
   } else if(date_format == "M"){
     gsub("-", "", substr(as.character(x), 1, 7))
-  } else if(date_format == "A"){
-    substr(as.character(x), 1, 4)
-  } else {
-    stop("Unknown date format.", call. = FALSE)
+  } else if(date_format %in% c("W", "D")){
+    gsub("-", "", as.character(x))
   }
 }
 
 #' @export
 #' @rdname eiadate
-eiadate_to_date_seq <- function(start, end){
-  date_format <- eiadate_format(start)
+eiadate_to_date_seq <- function(start, end, weekly = FALSE){
+  date_format <- eiadate_format(start, weekly)
   x <- eiadate_to_date(c(start, end))
   i <- switch(date_format,
               "Q" = "quarter", "M" = "month", "A" = "year",
@@ -57,19 +62,23 @@ eiadate_to_date_seq <- function(start, end){
 
 is_eiadate <- function(x){
   if(!is.character(x)) return(FALSE)
-  (grepl("^\\d\\d\\d\\d((0[1-9]|1[0-2])|)$", x)) |
-    grepl("^\\d\\d\\d\\d(Q|q)(1|2|3|4)$", x)
+  grepl("^\\d\\d\\d\\d((0[1-9]|1[0-2])|)$", x) |
+    grepl("^\\d\\d\\d\\dQ(1|2|3|4)$", x) |
+    grepl("^\\d\\d\\d\\d\\d\\d\\d\\d$", x)
 }
 
-eiadate_format <- function(x){
+eiadate_format <- function(x, weekly = FALSE, override = NULL){
+  if(!is.null(override)) return(override)
   if(any(!is_eiadate(x))) stop("Not an EIA format date string.", call. = FALSE)
-  if(grepl("Q|q", x[1])){
-    "Q"
-  } else if(grepl("^\\d\\d\\d\\d$", x[1])) {
-    "A"
+  if(grepl("^\\d\\d\\d\\d\\d\\d\\d\\d$", x[1])){
+    if(weekly) "W" else "D"
   } else if(grepl("^\\d\\d\\d\\d\\d\\d$", x[1])){
     "M"
+  } else if(grepl("^\\d\\d\\d\\d$", x[1])){
+    "A"
+  } else if(grepl("Q", x[1])){
+    "Q"
   } else {
-    print("Unhandled case")
+    stop("Unknown date format.", call. = FALSE)
   }
 }